Reset refactorings
